+++ src/core/config/config.interfaces.ts
@@ -116,6 +116,7 @@ export interface HomebridgeUiConfig {
   scheduledBackupPath?: string
   scheduledBackupDisable?: boolean
   disableServerMetricsMonitoring?: boolean
+  enableMdnsAdvertise?: boolean
   terminal?: {
     persistence?: boolean
     hideWarning?: boolean