+++ ui/src/app/modules/settings/settings.component.ts
@@ -186,6 +186,9 @@ export class SettingsComponent implements OnInit {
   public hbMDnsIsSaving = false
   public hbMDnsFormControl = new FormControl('')
 
+  public enableMdnsAdvertiseFormControl = new FormControl(false)
+  public enableMdnsAdvertiseIsSaving = false
+
   public hbPortIsInvalid = false
   public hbPortIsSaving = false
   public hbPortFormControl = new FormControl(0)
@@ -633,6 +636,11 @@ export class SettingsComponent implements OnInit {
     this.uiMetricsFormControl.valueChanges
       .pipe(debounceTime(750))
       .subscribe((value: boolean) => this.uiMetricsSave(value))
+
+    this.enableMdnsAdvertiseFormControl.patchValue(this.$settings.env.enableMdnsAdvertise || false)
+    this.enableMdnsAdvertiseFormControl.valueChanges
+      .pipe(debounceTime(750))
+      .subscribe((value: boolean) => this.enableMdnsAdvertiseSave(value))
 
     this.uiAccDebugFormControl.patchValue(this.$settings.env.accessoryControl?.debug)
     this.uiAccDebugFormControl.valueChanges
@@ -1562,6 +1570,27 @@ export class SettingsComponent implements OnInit {
       console.error(error)
       this.$toastr.error(error.message, this.$translate.instant('toast.title_error'))
       this.uiMetricsIsSaving = false
+    }
+  }
+
+  private async enableMdnsAdvertiseSave(value: boolean) {
+    try {
+      this.enableMdnsAdvertiseIsSaving = true
+      this.$settings.setEnvItem('enableMdnsAdvertise', value)
+      await this.saveUiSettingChange('enableMdnsAdvertise', value)
+      setTimeout(() => {
+        this.enableMdnsAdvertiseIsSaving = false
+        this.$api.put('/platform-tools/hb-service/set-full-service-restart-flag', {}).subscribe({
+          next: () => this.showRestartToast(),
+          error: (error) => {
+            console.error(error)
+            this.showRestartToast()
+          },
+        })
+      }, 1000)
+    } catch (error) {
+      console.error(error)
+      this.$toastr.error(error.message, this.$translate.instant('toast.title_error'))
     }
   }